Back/feature/7 five minutes

+++ src/main/java/com/parkinglot/service/PushService.java
@@ -5,13 +5,17 @@ import com.parkinglot.fcm.FcmUtil;
 import com.parkinglot.repository.ParkingRepository;
 import com.parkinglot.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
+@Slf4j
 public class PushService {
     private final ParkingRepository parkingRepository;
     private final ParkingService parkingService;
@@ -23,39 +27,41 @@ public class PushService {
         List<User> allUser = userRepository.findAll();
         for (User user : allUser) {
             if(user.getDeviceToken() != null){
-                String deviceToken = user.getDeviceToken();
-                Platform platform = user.getPlatform();
-                PushStatus pushStatus = user.getPushStatus();
-
                 List<PushDetail> pushDetails = user.getPushDetails();
 
-                if(pushStatus == PushStatus.ENABLE_ONLY) {
-                    for (PushDetail pushDetail : pushDetails) {
-                        String sector = pushDetail.getSector();
-                        Parking parking = parkingRepository.findBySector(sector).get();
-                        Long parkingId = parking.getId();
-                        int seat = pushDetail.getSeat();
+                if(!pushDetails.isEmpty()) {
+                    String deviceToken = user.getDeviceToken();
+                    Platform platform = user.getPlatform();
+                    PushStatus pushStatus = user.getPushStatus();
 
-                        parkingService.updateEnableDeviceToken(parkingId, seat, deviceToken, platform);
-                    }
-                } else if(pushStatus == PushStatus.DISABLE_ONLY) {
-                    for (PushDetail pushDetail : pushDetails) {
-                        String sector = pushDetail.getSector();
-                        Parking parking = parkingRepository.findBySector(sector).get();
-                        Long parkingId = parking.getId();
-                        int seat = pushDetail.getSeat();
+                    if (pushStatus == PushStatus.ENABLE_ONLY) {
+                        for (PushDetail pushDetail : pushDetails) {
+                            String sector = pushDetail.getSector();
+                            Parking parking = parkingRepository.findBySector(sector).get();
+                            Long parkingId = parking.getId();
+                            int seat = pushDetail.getSeat();
 
-                        parkingService.updateDisableDeviceToken(parkingId, seat, deviceToken, platform);
-                    }
-                } else {        //pushStatus.BOTH 일때
-                    for (PushDetail pushDetail : pushDetails) {
-                        String sector = pushDetail.getSector();
-                        Parking parking = parkingRepository.findBySector(sector).get();
-                        Long parkingId = parking.getId();
-                        int seat = pushDetail.getSeat();
+                            parkingService.updateEnableDeviceToken(parkingId, seat, deviceToken, platform);
+                        }
+                    } else if (pushStatus == PushStatus.DISABLE_ONLY) {
+                        for (PushDetail pushDetail : pushDetails) {
+                            String sector = pushDetail.getSector();
+                            Parking parking = parkingRepository.findBySector(sector).get();
+                            Long parkingId = parking.getId();
+                            int seat = pushDetail.getSeat();
 
-                        parkingService.updateEnableDeviceToken(parkingId, seat, deviceToken, platform);
-                        parkingService.updateDisableDeviceToken(parkingId, seat, deviceToken, platform);
+                            parkingService.updateDisableDeviceToken(parkingId, seat, deviceToken, platform);
+                        }
+                    } else if (pushStatus == PushStatus.BOTH) {        //pushStatus.BOTH 일때
+                        for (PushDetail pushDetail : pushDetails) {
+                            String sector = pushDetail.getSector();
+                            Parking parking = parkingRepository.findBySector(sector).get();
+                            Long parkingId = parking.getId();
+                            int seat = pushDetail.getSeat();
+
+                            parkingService.updateEnableDeviceToken(parkingId, seat, deviceToken, platform);
+                            parkingService.updateDisableDeviceToken(parkingId, seat, deviceToken, platform);
+                        }
                     }
                 }
             }
@@ -76,13 +82,14 @@ public class PushService {
         }else if (pushStatus == PushStatus.DISABLE_ONLY) {
             parkingService.updateDisableDeviceToken(parking.getId(), seat, deviceToken, platform);
             parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
-        }else {
+        }else if (pushStatus == PushStatus.BOTH) {
             parkingService.updateEnableDeviceToken(parking.getId(), seat, deviceToken, platform);
             parkingService.updateDisableDeviceToken(parking.getId(), seat, deviceToken, platform);
+        }else {
+            parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
+            parkingService.deleteDisableDeviceToken(parking.getId(), seat, deviceToken);
         }
 
-        user.setPushStatus(pushStatus);
-        userRepository.save(user);
     }
 
     //유저 아이디로 찾아서 푸쉬 삭제
@@ -96,13 +103,11 @@ public class PushService {
             parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
         }else if (pushStatus == PushStatus.DISABLE_ONLY) {
             parkingService.deleteDisableDeviceToken(parking.getId(), seat, deviceToken);
-        }else {
+        }else if (pushStatus == PushStatus.BOTH) {
             parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
             parkingService.deleteDisableDeviceToken(parking.getId(), seat, deviceToken);
         }
 
-        user.setPushStatus(pushStatus);
-        userRepository.save(user);
     }
 
     //유저 이름으로 찾아서 푸쉬 업데이트
@@ -119,13 +124,14 @@ public class PushService {
         }else if (pushStatus == PushStatus.DISABLE_ONLY) {
             parkingService.updateDisableDeviceToken(parking.getId(), seat, deviceToken, platform);
             parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
-        }else {
+        }else if (pushStatus == PushStatus.BOTH) {
             parkingService.updateEnableDeviceToken(parking.getId(), seat, deviceToken, platform);
             parkingService.updateDisableDeviceToken(parking.getId(), seat, deviceToken, platform);
+        }else {
+            parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
+            parkingService.deleteDisableDeviceToken(parking.getId(), seat, deviceToken);
         }
 
-        user.setPushStatus(pushStatus);
-        userRepository.save(user);
     }
 
     //유저 이름으로 찾아서 푸쉬 삭제
@@ -139,36 +145,42 @@ public class PushService {
             parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
         }else if (pushStatus == PushStatus.DISABLE_ONLY) {
             parkingService.deleteDisableDeviceToken(parking.getId(), seat, deviceToken);
-        }else {
+        }else if (pushStatus == PushStatus.BOTH) {
             parkingService.deleteEnableDeviceToken(parking.getId(), seat, deviceToken);
             parkingService.deleteDisableDeviceToken(parking.getId(), seat, deviceToken);
         }
 
-        user.setPushStatus(pushStatus);
-        userRepository.save(user);
     }
 
     //자리가 비었을 때 (enable) 푸쉬보내기
     public void sendEnablePush(String sector, int seat){
+        log.debug("----------[START] PushService - sendEnablePush----------");
         String message = sector + "섹터의 " + seat + "번 자리가 비었습니다.";
         Parking parking = parkingRepository.findBySector(sector).get();
         Map<String, Platform> deviceTokenMap = parkingService.getEnableDeviceTokenMap(parking.getId(), seat);
+        log.debug("디바이스 토큰 맵 목록 : {}",deviceTokenMap);
+        log.debug("맵이 비었는가? : {}",deviceTokenMap.isEmpty());
 
         for (String deviceToken : deviceTokenMap.keySet()) {
             Platform platform = deviceTokenMap.get(deviceToken);
             fcmUtil.sendFcm(deviceToken, "아몰랑파킹 주차알림", message, platform);
         }
+        log.debug("----------[END] PushService - sendEnablePush----------");
     }
 
     //자리가 찼을 떄 (disable) 푸쉬보내기
     public void sendDisablePush(String sector, int seat){
+        log.debug("----------[START] PushService - sendDisablePush----------");
         String message = sector + "섹터의 " + seat + "번 자리가 찼습니다.";
         Parking parking = parkingRepository.findBySector(sector).get();
         Map<String, Platform> deviceTokenMap = parkingService.getDisableDeviceTokenMap(parking.getId(), seat);
+        log.debug("디바이스 토큰 맵 목록 : {}",deviceTokenMap);
+        log.debug("맵이 비었는가? : {}",deviceTokenMap.isEmpty());
 
         for (String deviceToken : deviceTokenMap.keySet()) {
             Platform platform = deviceTokenMap.get(deviceToken);
             fcmUtil.sendFcm(deviceToken, "아몰랑파킹 주차알림", message, platform);
         }
+        log.debug("----------[END] PushService - sendDisablePush----------");
     }
 }
